requirements.txt
# Core LLM & Agent
openai>=1.50.0
google-generativeai>=0.8.0
transformers>=4.40.0
accelerate>=0.30.0

# Vision & Image Processing
torch>=2.2.0
torchvision>=0.17.0
Pillow>=10.0.0
opencv-python>=4.9.0
albumentations>=1.4.0
timm>=1.0.0

# CLIP Models
open_clip_torch>=2.24.0

# Vector Store & RAG
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0
langchain>=0.2.0
langchain-community>=0.2.0

# API & Services
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
requests>=2.32.0

# OCR & Text
pytesseract>=0.3.10
easyocr>=1.7.0

# Data & Storage
numpy>=1.26.0
pandas>=2.2.0
h5py>=3.11.0
PyYAML>=6.0.1
pydantic>=2.7.0
pydantic-settings>=2.3.0

# Grad-CAM
grad-cam>=1.5.0
matplotlib>=3.9.0
seaborn>=0.13.0

# Search & Web
duckduckgo-search>=6.1.0
arxiv>=2.1.0
beautifulsoup4>=4.12.0

# Segmentation
segment-anything>=1.0.0

# Utilities
tqdm>=4.66.0
loguru>=0.7.0
python-dotenv>=1.0.0
tenacity>=8.3.0
rich>=13.7.0

# Testing
pytest>=8.2.0
pytest-asyncio>=0.23.0

# Notebooks
jupyter>=1.0.0
ipywidgets>=8.1.0
